requirements.txt
python-telegram-bot==20.7
anthropic==0.45.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
aiohttp==3.9.1
cryptography==41.0.7
alembic==1.13.1  # для миграций БД
psycopg2-binary==2.9.9  # если используете PostgreSQL
python-jose==3.3.0  # для работы с JWT-токенами (если будете делать авторизацию)

httpx==0.25.2
pytz
